requirements.txt
tqdm
nltk
pandas
boto3
requests
regex
seqeval
psutil
pytorch>=1.9.1